+++ Back/src/main/java/julien/filrouge/forme/mesformes/Rond.java
@@ -4,7 +4,7 @@ public class Rond extends Shape {
 
     private double radius;
 
-    public Rond(Long id, String color, double x, double y, int radius) {
+    public Rond(Long id, String color, double x, double y, double radius) {
 
         super(id, color, x, y);
         this.radius = radius;
